grc: Factor out get_cmake_nproc and get_modifier_key
These functions are GUI-independent, and should therefore go into
a non-widget-specific location.
