+++ grc/gui_qt/components/executor.py
@@ -7,8 +7,8 @@ from pathlib import Path
 from shutil import which as find_executable
 from gnuradio import gr
 
-from ..Utils import get_cmake_nproc
 from ...core import Messages
+from ...core.utils.system import get_cmake_nproc
 
 
 class ExecFlowGraphThread(threading.Thread):
